requirements.txt
asgiref==3.2.10
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
idna==2.9
ImageHash==4.1.0
numpy==1.18.5
Pillow==7.1.2
pytz==2020.1
PyWavelets==1.1.1
requests==2.24.0
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
